backend/requirements.txt
fastapi==0.108.0
uvicorn[standard]==0.25.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
azure-storage-blob==12.19.0
azure-search-documents==11.4.0
azure-identity==1.15.0
openai==1.6.0
pillow==10.2.0
numpy>=1.26.0
pandas>=2.1.4
requests==2.31.0
aiofiles==23.2.0
python-json-logger==2.0.7
httpx==0.26.0
tenacity==8.2.3
aiohttp==3.9.1 